moved searchForm into its own component and fixed issue with state dropdown.
