+++ client/src/css/App.css
@@ -8,7 +8,6 @@
 .frosted-glass {
   background: rgba(255, 255, 255, 0.2);
   backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
   border-radius: 10px;
   border: 1px solid rgba(255, 255, 255, 0.3);
   padding: 2rem;
