+++ client/src/css/searchForm.css
@@ -1,0 +1,4 @@
+.dropdown-menu {
+    position: absolute;
+    z-index: 1000;
+  }
