--- client/src/interfaces/coordinates.tsx
@@ -1,4 +1,0 @@
-export interface Coordinates {
-  lat: string;
-  lon: string;
-}
